Fix buggy persistent window-scroll position
Before, we set up storing the scroll position as a separate instruction,
similar to setting up 'load' events.

Normally, the 'scroll' event could fire before 'onload'. For example,
the browser might scroll after refreshing the page (the page is mostly
empty), save this position which is near the top, then load the content.
When we scrolled down programmatically, the stored scroll position would
be the one near the top, not the one we saved before refreshing (at least I
think that's what was happening).

To solve this, we can start saving the scroll position right after
placing the threads on the page.